requirements.txt
streamlit==1.33.0
pandas==2.2.1
matplotlib==3.8.4
requests==2.31.0
beautifulsoup4==4.12.3
reportlab==4.1.0
google-api-python-client==2.122.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
pybase64==1.2.1
lxml==5.2.1
python-dotenv==1.0.1

